moved enter student logic to seperate component
