+++ src/App.js
@@ -2,6 +2,7 @@ import "./App.css";
 import { useState } from "react";
 import BaseInput from "./BaseInput";
 import BestCombination from "./BestCombination";
+import EnterStudent from "./EnterStudent";
 
 const testData = [
   {
@@ -52,21 +53,11 @@ function App() {
   // enter student data logic
   const [enteredStudents, setEnteredStudents] = useState([]);
 
-  const [studentName, setStudentName] = useState("");
-  const [studentEarnings, setStudentEarnings] = useState("");
-  const [studentHours, setStudentHours] = useState("");
-
-  const handleNamechange = (string) => {
-    setStudentName(string);
-  };
-  const handleEarningschange = (num) => {
-    setStudentEarnings(num);
-  };
-  const handleHourschange = (num) => {
-    setStudentHours(num);
-  };
-
-  const handleSubmitStudent = () => {
+  const handleSubmitedStudent = (
+    studentName,
+    studentEarnings,
+    studentHours
+  ) => {
     const student = {
       name: studentName,
       earnings: parseInt(studentEarnings),
@@ -80,9 +71,6 @@ function App() {
       }
       return newStudentArray;
     });
-    setStudentName("");
-    setStudentEarnings("");
-    setStudentHours("");
   };
 
   const handleClearData = () => {
@@ -107,27 +95,11 @@ function App() {
       />
       <button onClick={handleProceede}>Proceede to enter Data</button>
       {enteringData && (
-        <>
-          <p>
-            entering student {enteredStudents.length + 1} / {numOfStudents}
-          </p>
-          <BaseInput
-            placeholder="Name"
-            inputChanged={handleNamechange}
-            inputValue={studentName}
-          />
-          <BaseInput
-            placeholder="Earnings Potential"
-            inputChanged={handleEarningschange}
-            inputValue={studentEarnings}
-          />
-          <BaseInput
-            placeholder="Instruction Hours Needed"
-            inputChanged={handleHourschange}
-            inputValue={studentHours}
-          />
-          <button onClick={handleSubmitStudent}>Submit Student</button>
-        </>
+        <EnterStudent
+          handleSubmitedStudent={handleSubmitedStudent}
+          enteredStudents={enteredStudents}
+          numOfStudents={numOfStudents}
+        />
       )}
       {enteredStudents && (
         <>
